pyproject.toml
[project]
name = "chatbot-medical-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.5.6",
    "google-genai>=1.18.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pypdf>=5.5.0",
    "ruff>=0.11.12",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
    "weaviate-client>=4.14.4",
]
